refactor: flatten subpackages
