+++ alphaevolve/engine.py
@@ -35,9 +35,7 @@ class AlphaEvolve:
         self.initial_program_paths = [Path(p) for p in initial_program_paths]
         self.config_path = Path(config_path)
         self.store = store or ProgramStore()
-        self.controller = Controller(
-            self.store, initial_program_paths=self.initial_program_paths
-        )
+        self.controller = Controller(self.store, initial_program_paths=self.initial_program_paths)
 
     async def run(self, iterations: int = 1) -> Strategy:
         """Run the evolution loop for a fixed number of iterations."""
